requirements.txt
streamlit==1.38.0
google-generativeai==0.8.3
python-dotenv==0.21.0